Adding an id to the user table
The reason is that we need some unique identifier in the API URLs. Using the email is awkward.
Furthermore, this identifier will need to be stored in the Top Cluck side, and it's much more
normal to be storing some number/guid of a third-party service than it is to save something like
the "COOP email".